Remove legacy sync-extract function
